% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jptemplate.R
\name{create_project}
\alias{create_project}
\title{Insert R Project Template}
\usage{
create_project(
  title,
  subtitle = NULL,
  gitignore = FALSE,
  readme = TRUE,
  rmarkdown = TRUE,
  folders = TRUE
)
}
\description{
This function will insert the R project template into the current R project template folder
}
\details{
* title        = 'the name of the project' \n
* subtitle     = 'the subtitle of the project' \n
* gitignore    = Set to TRUE if you want to ignore the gitignore file \n
* readme     = FALSE disables the creation of a template readme file \n
* rmarkdown  = FALSE disables the creation of a template rmarkdown file \n
* folders    = FALSE disables the creation of folders \n

It's callback is at: inst/rstudio/templates/project/create_project.dcf
}
