% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jptemplate.R
\name{create_project}
\alias{create_project}
\title{Insert R Project Template}
\usage{
create_project(
  title,
  subtitle = NULL,
  author = NULL,
  gitignore = FALSE,
  readme = TRUE,
  rmarkdown = TRUE,
  folders = TRUE
)
}
\arguments{
\item{title}{a mandatory character input of 'the name of the project'.}

\item{subtitle}{an optional character input of 'the subtitle of the project'.}

\item{author}{an optional character input of 'the project author'.}

\item{gitignore}{logical. Set to \code{TRUE} if you want to ignore the \code{.gitignore} file.}

\item{readme}{logical. \code{FALSE} disables the creation of a template \code{README.md} file.}

\item{rmarkdown}{logical. \code{FALSE} disables the creation of a template \code{.Rmd} file.}

\item{folders}{logical. \code{FALSE} disables the creation of folders.}
}
\description{
This function will insert the R project template into the current R project
template folder including a default set of sub-folders, a \code{README.md} template file,
 and the default \code{.Rmd} file. All files will automatically contain the project name
 used in the function.
}
\examples{
create_project("Project_Name", subtitle = "Project_Subtitle", gitignore = TRUE, rmarkdown = FALSE)

}
